10.03.2020

+++ controllers/delete.js
@@ -1,8 +1,15 @@
 export function deleteResource(req, res) {
     const resource = res.locals.resource;
-    resource.setPredicateToDelete(req.params.attributeName, req.body.value);
-    resource
-        .delete()
-        .then(data => res.send({ status: true }))
-        .catch(err => res.status(500).send({ status: false, msg: err }));
+    if (req.params.attributeName != undefined) {
+        resource.setPredicateToDelete(req.params.attributeName, req.body.value);
+        resource
+            .delete()
+            .then(data => res.send({ status: true }))
+            .catch(err => res.status(500).send({ status: false, msg: err }));
+    } else {
+        resource
+            .completeDelete()
+            .then(() => res.send({ status: true }))
+            .catch(err => res.status(500).send({ status: false, msg: err }));
+    }
 }
